+++ upload/manage_photos.php
@@ -1,7 +1,10 @@
 <?php
-define("THIS_PAGE",'manage_photos');
-define("PARENT_PAGE","photos");
+define('THIS_PAGE','manage_photos');
+define('PARENT_PAGE','photos');
 require 'includes/config.inc.php';
+
+global $userquery,$cbphoto,$eh,$pages,$db;
+
 $userquery->logincheck();
 $udetails = $userquery->get_user_details(userid());
 assign('user',$udetails);
@@ -12,14 +15,19 @@ $mode = $_GET['mode'];
 $page = mysql_clean($_GET['page']);
 $get_limit = create_query_limit($page,MAINPLIST);
 
-assign('queryString',queryString(NULL,array('type',
-					'makeProfileItem',
-					'removeProfileItem',
-					'delete_photo')));
+assign(
+'queryString',
+    queryString(NULL,[
+        'type',
+        'makeProfileItem',
+        'removeProfileItem',
+        'delete_photo']
+    )
+);
 
 switch($mode)
 {
-	case "uploaded":
+	case 'uploaded':
 	default:
 		assign('mode','uploaded');
 		if(isset($_GET['delete_photo'])) {
@@ -27,13 +35,13 @@ switch($mode)
 			$cbphoto->delete_photo($id);	
 		}
 		
-		if(isset($_POST['delete_photos'])) {
+		if(isset($_POST['delete_photos']) && is_array($_POST['check_photo'])) {
 			$total = count($_POST['check_photo']);
 			for($i=0;$i<$total;$i++) {
 				$cbphoto->delete_photo($_POST['check_photo'][$i]);	
 			}
 			$eh->flush();
-			e(sprintf(lang("total_photos_deleted"),$total),"m");
+			e(sprintf(lang('total_photos_deleted'),$total),'m');
 		}
 		
 		//Setting Profile Photo
@@ -48,7 +56,11 @@ switch($mode)
 			$userquery->removeProfileItem();
 		}
 		
-		$photo_arr = array("user"=>userid(),"limit"=>$get_limit, 'order'=>' date_added DESC');
+		$photo_arr = [
+            'user'  => userid(),
+            'limit' => $get_limit,
+            'order' => ' date_added DESC'
+        ];
 		
 		if(get('query') != '') {
 			$photo_arr['title'] = mysql_clean(get('query'));
@@ -65,10 +77,10 @@ switch($mode)
 		
 		//Pagination
 		$pages->paginate($total_pages,$page);
-		subtitle(lang("manage_photos"));
+		subtitle(lang('manage_photos'));
 	    break;
 	
-	case "favorite":
+	case 'favorite':
 		assign('mode','favorite');
 		if($_GET['remove_fav_photo']) {
 			$photo = mysql_clean($_GET['remove_fav_photo']);
@@ -76,22 +88,21 @@ switch($mode)
 			updateObjectStats('fav','photo',$photo,'-');	
 		}
 		
-		if($_POST['remove_fav_photos']) {
+		if($_POST['remove_fav_photos'] && is_array($_POST['check_photo'])) {
 			$total = count($_POST['check_photo']);
-			for($i=0;$i<$total;$i++)
-			{
+			for($i=0;$i<$total;$i++) {
 				$cbphoto->action->remove_favorite($_POST['check_photo'][$i]);
 				updateObjectStats('fav','photo',$_POST['check_photo'][$i],'-');	
 			}
 			$eh->flush();
-			e(sprintf(lang("total_fav_photos_removed"),$total),"m");
+			e(sprintf(lang('total_fav_photos_removed'),$total),'m');
 		}
 		
 		if(get('query')!='') {
-			$cond = " (photos.photo_title LIKE '%".mysql_clean(get('query'))."%' OR photos.photo_tags LIKE '%".mysql_clean(get('query'))."%' )";
+			$cond = ' (photos.photo_title LIKE \'%'.mysql_clean(get('query')).'%\' OR photos.photo_tags LIKE \'%'.mysql_clean(get('query')).'%\' )';
 		}
 		
-		$photo_arr = array('user'=>userid(),"limit"=>$get_limit,"cond"=>$cond);
+		$photo_arr = ['user'=>userid(),'limit'=>$get_limit,'cond'=>$cond];
 		$photos = $cbphoto->action->get_favorites($photo_arr);
 		assign('photos',$photos);
 		
@@ -101,17 +112,17 @@ switch($mode)
 		
 		//Pagination
 		$pages->paginate($total_pages,$page);
-		subtitle(lang("manage_favorite_photos"));
+		subtitle(lang('manage_favorite_photos'));
 	    break;
 	
-	case "my_album":
+	case 'my_album':
 		assign('albumPrivacyUrl',queryString('','album_privacy'));
 		assign('mode','orphan');
 		
 		if(isset($_GET['album_privacy'])) {
-			if(in_array(get('album_privacy'),array('private','public','friends'))) {
-				$db->update(tbl("users"),array("album_privacy"),array(mysql_clean(get("album_privacy")))," userid='".userid()."'" );
-				e(lang("album_privacy_updated"),'m');
+			if(in_array(get('album_privacy'),['private','public','friends'])) {
+				$db->update(tbl('users'),['album_privacy'],[mysql_clean(get('album_privacy'))],' userid=\''.userid().'\'' );
+				e(lang('album_privacy_updated'),'m');
 				$udetails ['album_privacy'] = get('album_privacy');
 				assign('user',$udetails);
 			}
@@ -122,19 +133,18 @@ switch($mode)
 			$cbphoto->delete_photo($id);		
 		}
 		
-		if(isset($_POST['delete_orphan_photos'])) {
+		if(isset($_POST['delete_orphan_photos']) && is_array($_POST['check_photo'])) {
 			$total = count($_POST['check_photo']);
 			for($i=0;$i<$total;$i++) {
 				$cbphoto->delete_photo($_POST['check_photo'][$i],TRUE);
 			}
 			$eh->flush();
-			e(sprintf(lang("total_photos_deleted"),$total),"m");
+			e(sprintf(lang('total_photos_deleted'),$total),'m');
 		}
-		$photo_arr = array("user"=>userid(),"limit"=>$get_limit, 'order'=>' date_added DESC', "get_orphans"=>TRUE);
-		$collection = $cbphoto->collection->get_collections(array("user"=>userid(),"type"=>"photos"));
+		$photo_arr = ['user'=>userid(),'limit'=>$get_limit, 'order'=>' date_added DESC', 'get_orphans'=>TRUE];
+		$collection = $cbphoto->collection->get_collections(['user'=>userid(),'type'=>'photos']);
 		
-		if(get('query') != '')
-		{
+		if(get('query') != '') {
 			$photo_arr['title'] = mysql_clean(get('query'));
 			$photo_arr['tags']	= mysql_clean(get('query'));
 		}
@@ -149,7 +159,7 @@ switch($mode)
 		
 		//Pagination
 		$pages->paginate($total_pages,$page);
-		subtitle(lang("manage_orphan_photos"));
+		subtitle(lang('manage_orphan_photos'));
 	    break;
 }
 template_files('manage_photos.html');
